+++ src/main/java/nc/apps/smartadder/dao/interfaces/LanguageDAO.java
@@ -4,7 +4,7 @@ import nc.apps.smartadder.dao.exception.DAOException;
 import nc.apps.smartadder.domain.Language;
 
 public interface LanguageDAO {
-    Long save(Language language) throws DAOException;
-    Long getIdByValues(Language language) throws DAOException;
+    Integer save(Language language) throws DAOException;
+    Integer getIdByValues(Language language) throws DAOException;
 
 }
